chat server 支持多协议定义
