+++ chatConsole/chatServer/src/main/java/com/bytrees/chat/ChatChannelHandlerFactory.java
@@ -1,0 +1,25 @@
+package com.bytrees.chat;
+
+import com.bytrees.chat.channelhandler.ProtobufChannelHandler;
+
+import io.netty.channel.ChannelHandler;
+import io.netty.channel.ChannelInitializer;
+import io.netty.channel.socket.SocketChannel;
+
+public class ChatChannelHandlerFactory {
+	private ChatChannelHandlerFactory() {}
+
+	public static ChannelHandler getChannelInitializer(ChatProtocolEnum chatProtocol) {
+		if (chatProtocol.equals(ChatProtocolEnum.STRINGLINE)) {
+			
+		} else if (chatProtocol.equals(ChatProtocolEnum.PROTOBUF)) {
+			return new ChannelInitializer<SocketChannel>() {
+				@Override
+				protected void initChannel(SocketChannel ch) throws Exception {
+					ch.pipeline().addLast(new ProtobufChannelHandler());
+				}
+			};
+		}
+		return null;
+	}
+}
