+++ chatConsole/chatServer/src/main/java/com/bytrees/chat/ChatProtocolEnum.java
@@ -1,0 +1,14 @@
+package com.bytrees.chat;
+
+public enum ChatProtocolEnum {
+	STRINGLINE(1),
+	PROTOBUF(2);
+
+	private int typeId;
+	private ChatProtocolEnum(int typeId) {
+		this.typeId = typeId;
+	}
+	public int getTypeId() {
+		return typeId;
+	}
+}
